+++ dev/launch_epr_service_storage_indexer.sh
@@ -11,7 +11,7 @@ CURRENT_DIR="$(pwd)"
 SCRIPT_DIR="$( cd -- "$(dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 
 usage() {
-    echo "$0 [-b <bucket_name>] [-p <epr_address>] [-e <emulator_address>] [-i <index_path>] [-c <config_path>] [-s] [-C] [-h]"
+    echo "$0 [-b <bucket_name>] [-p <epr_address>] [-e <emulator_address>] [-i <index_path>] [-c <config_path>] [-s] [-C] [-d] [-h]"
     echo -e "\t-b <bucket_name>: Bucket name. Default: example"
     echo -e "\t-p <epr_address>: Address of the package registry service. Default: localhost:8080"
     echo -e "\t-e <emulator_address>: Address of the emulator host (fake GCS server). Default: localhost:4443"
@@ -20,6 +20,7 @@ usage() {
     echo -e "\t-c <config_path>: Path to the configurastion file. Default: \"\""
     echo -e "\t-s : Enable SQL Storage indexer. By default Storage Indexer is enabled."
     echo -e "\t-C : Enable Search Cache. Just supported with SQL Storage indexer. By default Search Cache is disabled."
+    echo -e "\t-d: Enable debug mode. Default: false"
     echo -e "\t-h: Show this message"
 }
 
@@ -30,8 +31,9 @@ EMULATOR_HOST="localhost:4443"
 CONFIG_PATH="${SCRIPT_DIR}/../config.yml"
 ENABLE_STORAGE_SQL_INDEXER=0
 ENABLE_SEARCH_CACHE=0
+ENABLE_DEBUG_MODE=0
 
-while getopts ":b:p:i:e:c:sCh" o; do
+while getopts ":b:p:i:e:c:sdCh" o; do
   case "${o}" in
     b)
       BUCKET_NAME="${OPTARG}"
@@ -53,6 +55,9 @@ while getopts ":b:p:i:e:c:sCh" o; do
       ;;
     C)
       ENABLE_SEARCH_CACHE=1
+      ;;
+    d)
+      ENABLE_DEBUG_MODE=1
       ;;
     h)
       usage
@@ -103,7 +108,9 @@ fi
 export EPR_DISABLE_PACKAGE_VALIDATION="true"
 export EPR_ADDRESS="${ADDRESS}"
 
-# export EPR_LOG_LEVEL="debug"
+if [[ "${ENABLE_DEBUG_MODE}" == 1 ]]; then
+    export EPR_LOG_LEVEL="debug"
+fi
 export EPR_CONFIG="${CONFIG_PATH}"
 # export EPR_SQL_INDEXER_DATABASE_FOLDER_PATH=/tmp
 # export EPR_SQL_INDEXER_SEARCH_CACHE_SIZE=100